Edit: - AuthorController > block delete author when he have book in db - BookController > change redirect addres - Book > refactor release_date variable name - BookService > refactor release_date variable name - bookList.jsp > refactor release_date variable name - bookAdd.jsp > refactor release_date variable name

+++ src/main/java/pl/dm/libraryprojectdm/controller/AuthorController.java
@@ -39,7 +39,7 @@ public class AuthorController {
 
     @PostMapping("/delete")
     public String deleteAuthor(@RequestParam Long authorId) {
-        if (authorId != null) {
+        if (authorId != null && authorService.findAuthorById(authorId).getBooks().size() == 0) {
             try {
                 authorService.delete(authorId);
                 return "redirect:/author/list";
@@ -47,6 +47,7 @@ public class AuthorController {
                 System.out.println("Error");
             }
         }
+        System.out.println("Error");
         return "redirect:/author/list";
     }
